requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.1
httptools==0.4.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
PyMySQL==1.0.2
pypika-tortoise==0.1.6
pytz==2022.2.1
requests==2.28.1
sanic==22.6.2
sanic-routing==22.3.0
soupsieve==2.3.2.post1
tortoise-orm==0.19.2
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvloop==0.16.0
websockets==10.3
yarl==1.8.1